try redirecting

+++ StockistTeam8/src/main/java/sg/edu/iss/team8/controller/AdminUserController.java
@@ -1,8 +1,12 @@
 package sg.edu.iss.team8.controller;
+
 //testing conflicts
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,7 +18,10 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.method.HandlerMethod;
 import org.springframework.web.servlet.ModelAndView;
+import org.springframework.web.servlet.mvc.method.RequestMappingInfo;
+import org.springframework.web.servlet.mvc.method.annotation.RequestMappingHandlerMapping;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import sg.edu.iss.team8.exception.UserNotFound;
@@ -45,8 +52,19 @@ public class AdminUserController {
 	 * @return
 	 */
 
+	@Autowired
+	private RequestMappingHandlerMapping requestMappingHandlerMapping;
+	
+	@RequestMapping(value = "/**", method = RequestMethod.GET)
+	public String firstTime(HttpServletRequest request, HttpSession session) {
+		Map<RequestMappingInfo, HandlerMethod> mapping = requestMappingHandlerMapping.getHandlerMethods();
+		return new TestController().testURL(request, session, mapping); 
+	}
+
 	@RequestMapping(value = "/create", method = RequestMethod.GET)
-	public ModelAndView newUserPage() {
+	public ModelAndView newUserPage(HttpSession session) {
+		if (!new TestController().isAdmin(session))
+			return new ModelAndView("403");
 		ModelAndView mav = new ModelAndView("user-new", "user", new User());
 		ArrayList<User> userList = uService.findAllUsers();
 		mav.addObject("userList", userList);
@@ -55,8 +73,10 @@ public class AdminUserController {
 
 	@RequestMapping(value = "/create", method = RequestMethod.POST)
 	public ModelAndView createNewUser(@ModelAttribute @Valid User user, BindingResult result,
-			final RedirectAttributes redirectAttributes) {
-
+			final RedirectAttributes redirectAttributes, HttpSession session) {
+		if (!new TestController().isAdmin(session))
+			return new ModelAndView("403");
+		
 		if (result.hasErrors())
 			return new ModelAndView("user-new");
 
@@ -68,10 +88,12 @@ public class AdminUserController {
 
 		redirectAttributes.addFlashAttribute("message", message);
 		return mav;
-	} 
+	}
 
 	@RequestMapping(value = "/list", method = RequestMethod.GET)
-	public ModelAndView userListPage() {
+	public ModelAndView userListPage(HttpSession session) {
+		if (!new TestController().isAdmin(session))
+			return new ModelAndView("403");
 		ModelAndView mav = new ModelAndView("manageuser");
 		List<User> userList = uService.findAllUsers();
 		mav.addObject("userList", userList);
@@ -79,7 +101,9 @@ public class AdminUserController {
 	}
 
 	@RequestMapping(value = "/edit/{id}", method = RequestMethod.GET)
-	public ModelAndView editUserPage(@PathVariable String id) {
+	public ModelAndView editUserPage(@PathVariable String id, HttpSession session) {
+		if (!new TestController().isAdmin(session))
+			return new ModelAndView("403");
 		ModelAndView mav = new ModelAndView("user-edit");
 		User user = uService.findUser(id);
 		mav.addObject("user", user);
@@ -90,7 +114,9 @@ public class AdminUserController {
 
 	@RequestMapping(value = "/edit/{id}", method = RequestMethod.POST)
 	public ModelAndView editUser(@ModelAttribute @Valid User user, BindingResult result, @PathVariable String id,
-			final RedirectAttributes redirectAttributes) throws UserNotFound {
+			final RedirectAttributes redirectAttributes, HttpSession session) throws UserNotFound {
+		if (!new TestController().isAdmin(session))
+			return new ModelAndView("403");
 
 		if (result.hasErrors())
 			return new ModelAndView("user-edit");
@@ -105,9 +131,10 @@ public class AdminUserController {
 	}
 
 	@RequestMapping(value = "/delete/{id}", method = RequestMethod.GET)
-	public ModelAndView deleteUser(@PathVariable String id, final RedirectAttributes redirectAttributes)
+	public ModelAndView deleteUser(@PathVariable String id, final RedirectAttributes redirectAttributes, HttpSession session)
 			throws UserNotFound {
-
+		if (!new TestController().isAdmin(session))
+			return new ModelAndView("403");
 		ModelAndView mav = new ModelAndView("redirect:/admin/user/list");
 		User user = uService.findUser(id);
 		uService.removeUser(user);
